+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 1.7.6
+- Carrier integration
+  - Fix: don't show error to user in UI when checking for `*.integration.json` existence on `create`/`update` 
+
+## 1.7.5
+- Carrier integration
+  - Fix: now checks correctly if carrier already exists on `create`
+  - Refactored some code for future maintainability
+
 ## 1.7.4
 - Carrier integration
   - Fix: directly load integration to dashboard when loading from file (without showing an 'empty' dashboard in between)